+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Unreleased
 
+### Changes
+
+* Add ActiveSupport 6.0 support
+  * [PR#8](https://github.com/yujideveloper/rekkyo/pull/8)
+
 
 ## 1.0.0 (2019-04-08)