+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Drop Ruby 2.4 support
   * [PR#14](https://github.com/yujideveloper/rekkyo/pull/14)
 * Call `super` when `Rekkyo::Type` is included
+* Add Ruby 3.1 and ActiveSupport 3.1 support
 
 ### Misc